// resources/js/api/generated/models/getApiV1AdminSecurityAuditLogsStats200Data.ts
/**
 * Generated by orval v8.0.2 🍺
 * Do not edit manually.
 * Berth API
 * Berth: Opinionated docker compose stack management API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiV1AdminSecurityAuditLogsStats200DataEventsByCategory } from './getApiV1AdminSecurityAuditLogsStats200DataEventsByCategory';
import type { GetApiV1AdminSecurityAuditLogsStats200DataEventsBySeverity } from './getApiV1AdminSecurityAuditLogsStats200DataEventsBySeverity';
import type { GetApiV1AdminSecurityAuditLogsStats200DataRecentEventTypesItem } from './getApiV1AdminSecurityAuditLogsStats200DataRecentEventTypesItem';

export type GetApiV1AdminSecurityAuditLogsStats200Data = {
  events_by_category: GetApiV1AdminSecurityAuditLogsStats200DataEventsByCategory;
  events_by_severity: GetApiV1AdminSecurityAuditLogsStats200DataEventsBySeverity;
  events_last_24_hours: number;
  events_last_7_days: number;
  failed_events: number;
  recent_event_types: GetApiV1AdminSecurityAuditLogsStats200DataRecentEventTypesItem[];
  total_events: number;
};